+++ gazebo/gui/ConfigWidget_TEST.hh
@@ -39,6 +39,9 @@ class ConfigWidget_TEST : public QTestFixture
   /// \brief Test creating a config widget from a visual message.
   private slots: void VisualMsgWidget();
 
+  /// \brief Test creating a config widget from a plugin message.
+  private slots: void PluginMsgWidget();
+
   /// \brief Test setting visibility of a field in config widget.
   private slots: void ConfigWidgetVisible();
 
@@ -48,6 +51,9 @@ class ConfigWidget_TEST : public QTestFixture
   /// \brief Test creating and updating a config widget without parsing
   /// messages.
   private slots: void CreatedExternally();
+
+  /// \brief Test functions related to enum config widgets.
+  private slots: void EnumConfigWidget();
 
   /// \brief Test receiving a signal from child uint widget.
   private slots: void ChildUIntSignal();